Recreate add function as a module
