+++ src/modules/tasks.js
@@ -31,6 +31,6 @@
 //   },
 // ];
 
-const toDos = [];
+// const toDos = [];
 
-export default toDos;
+// export default toDos;
